+++ python/htpclient/download.py
@@ -6,24 +6,26 @@ import sys
 import os
 
 from htpclient.initialize import Initialize
+from htpclient.session import Session
 
 
 class Download:
     @staticmethod
     def download(url, output, no_header=False):
         try:
+            session = Session().s
             if Initialize.get_os() == 1:
                 output = output.replace("/", '\\')
 
             # Check header
             if not no_header:
-                head = requests.head(url)
+                head = session.head(url)
                 # not sure if we only should allow 200/302, but then it's present for sure
                 if head.status_code != 200 and head.status_code != 302:
                     return False
 
             with open(output, "wb") as file:
-                response = requests.get(url, stream=True)
+                response = session.get(url, stream=True)
                 total_length = response.headers.get('Content-Length')
 
                 if total_length is None:  # no content length header