+++ python/changelog.md
@@ -3,6 +3,7 @@
 ### Enhancements
 
 * Added check for chunk length 0 sent from the server to avoid full agent crash on running.
+* Using requests session
 
 ## 0.1.8 -> 0.2.0